Create Hashing.md
Hashing is section 9.4 on book Page-324. In the first reading, I skip the illustration of the example code and performance analysis with equation derivation. I focus on the principle and advantage and disadvantage of the hashing and its implementations.
Hashing is used as an efficient solution to solve "Two Sum"  in Leetcode.